refactor(Invoice form): move component to its own file
